+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.1.3
 * Fix issue with page/paging parameter
 * Fix issue with limited number of parameters
+* Fix double encoding of parameters
 
 ## 0.1.2
 * Fix for Activities resources